requirements.txt
aiodns==2.0.0
asgiref==3.4.1
backcall==0.2.0
cffi==1.15.0
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.7
django-heroku==0.0.0
dnspython==1.15.0
dnspython3==1.15.0
Faker==8.12.2
gunicorn==20.1.0
idna==2.10
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
netaddr==0.8.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pycares==4.1.2
pycparser==2.21
pydnsbl==1.1.4
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
traitlets==5.1.0
wcwidth==0.2.5
whitenoise==5.3.0
